refactor: address PR review comments (round 2)
- Rename workflow to "Build SwiftPM" (tido64)
- Replace shallow clone + fetch with actions/checkout for Hermes (tido64)
- Use efficient git init + fetch pattern in buildFromHermesCommit (tido64)
- Extract reusable build env object in hermes.js (tido64)
- Move macOS version resolution to fork-only macosVersionResolver.js (tido64)
- Replace platformLinkerSettings with #if os(macOS) in Package.swift (Saad)
- Revert CMAKE_OSX_DEPLOYMENT_TARGET from build-apple-framework.sh (Saad)
- Add // [macOS] tag to os require in hermes.js (tido64)
